Add files via upload
This file contains an EDA on a dataset - The dataset contains 1.3 million Sarcastic comments from the Internet commentary website Reddit. The dataset was generated by scraping comments from Reddit (not by me) containing the \s ( sarcasm) tag. This tag is often used by Redditors to indicate that their comment is in jest and not meant to be taken seriously, and is generally a reliable indicator of sarcastic comment content.

The data was gathered by: Mikhail Khodak and Nikunj Saunshi and Kiran Vodrahalli for their article  "A Large Self-Annotated Corpus for Sarcasm".